Add license. Modify locations.
